Rewrite to use external `routePath` --> `annotations` storage object
This no longer uses Express' route objects to store annotations, instead
they are stored in a `routePath` --> `annotations` map located at:
`app.annotations`.

This enforces the idea that annotations are resource-based and they are
independent of the actual routes to get to those resources. With this
change, annotations are only available for string-based route paths.

The `findAll()` API has been modified to return a data structure that is
in the same format as `app.routes`. Calling `findAll()` can be thought
of as reducing the set of the app's `routes` down to those which match
the specified annotations.